+++ templates/ens/src/skills/check.ts
@@ -1,4 +1,5 @@
-import { Context, getUserInfo, Skill } from "@xmtp/message-kit";
+import { Context, Skill } from "@xmtp/message-kit";
+import { getUserInfo } from "xmtp";
 
 const ensUrl = "https://app.ens.domains/";
